Run github jobs using a docker-compose setup
Currently in github/gitlab we add build dependencies and/or
provide runtime configuration of the containers themselves
within the github/gitlab yaml definitions. As github actions
has far less support for templating/anchors, this results
in a significant amount of duplication.

Therefore, split runtime configuration of containers to
a docker-compose file, and build time configuration via
a Dockerfile. This then frees the github workflow to define
the remaining top level parameters via a matrix.

As a PoC demonstrate how this can be used for all existing tests,
and the no-cache overnight tests. This can be easily extended
to the overnight tests with cache present, and the randomised
tests as well.